Move {LoadOp, Operations, StoreOp} to wgt
